Update: GBHeaderTool Setup
GBHeaderTool seems to be working, at least for :window: added test reflection files for structs declared on their own, classes declared on their own, multiple structs declared in the same file, multiple classes declared in the same file and then a struct and a class declared in the same file.

Refactored code generator to make a bit more sense and added input/output folders as command arguments to better clarify where the generated files should end up!

Added function macros to search for (currently would do nothing but in the future I could try and automate something with them if I wanted to... - visual scripting would be a huge undertaking though and I would need to record the return type and the parameters). Validation as well to make sure the setup isn't borked straight away. Auto generates files based on header files that have a valid reflection setup (GB_REFLECT() and any GBVARIABLE/GBFUNCTION macros).

GBHeaderToolTest will run the GBHeaderTool on a given folder (Sandbox/Source/ReflectionTestFiles/AutoTestFolderToUse/) and output it to Sandbox/Source/GeneratedCode/. The test will pass if the reflection include is found (GBReflect.h) and if the reflection macros are present (GB_REFLECT_BEGIN, GB_REFLECT_MEMBER and GB_REFLECT_END). This will run for both a class and a struct. If either of these objects doesn't have any one of these macros, the test will fail.