Update GBClassRegistry.h
:penguin: / :apple: - Yeah you're right folks, it wasn't in the reflect namespace!

This is making me more keen on having some kind of offline validator to compile with WSL to try and catch any of these errors before they make it to the build machine... would be a cool thing to have.

Plus it would be good to use for the editor/engine project as well, not just the tools. I guess it just needs to take whatever the code input is and try and compile it on the WSL? Not sure how feasible that would be...