Update: Warnings/Error Window
Warnings/Errors will now be printed to the warning/errors window. User can scroll down the text box and copy the text to the clipboard (feedback with a tooltip to let them know they've copied something...). Seems to crash out when trying to compile a 2nd time with warnings/errors, see what's happening...