Update: WIP Vector Reflection
Updated logic to work out if a ItemType is a pointer, or smart ptr and perform the appropriate vector operations on that vector based on this. (E.g. if removing an item from a vector of pointers, free the memory before removing the item)

Need to figure out std::vector<std::weak_ptr>...