Update: WIP GBHeaderTool
Some header file include clean up.
Added GBBool which can be used as a work around for vector of bools since there is an added complexity with how they are stored in C++ standard. Intent is to just use the custom struct and draw a boolean
as the vector element instead of drawing a full class if the type descriptor is for this class...